+++ certora/harness/AaveTokenV3HarnessCommunity.sol
@@ -26,14 +26,14 @@ contract AaveTokenV3Harness is AaveTokenV3 {
 
   function getDelegatingProposition(address user) public view returns (bool) {
     return
-      _balances[user].delegationMode == uint8(DelegationMode.PROPOSITION_DELEGATED) ||
-      _balances[user].delegationMode == uint8(DelegationMode.FULL_POWER_DELEGATED);
+      _balances[user].delegationMode == DelegationMode.PROPOSITION_DELEGATED ||
+      _balances[user].delegationMode == DelegationMode.FULL_POWER_DELEGATED;
   }
 
   function getDelegatingVoting(address user) public view returns (bool) {
     return
-      _balances[user].delegationMode == uint8(DelegationMode.VOTING_DELEGATED) ||
-      _balances[user].delegationMode == uint8(DelegationMode.FULL_POWER_DELEGATED);
+      _balances[user].delegationMode == DelegationMode.VOTING_DELEGATED ||
+      _balances[user].delegationMode == DelegationMode.FULL_POWER_DELEGATED;
   }
 
   function getVotingDelegate(address user) public view returns (address) {
@@ -45,7 +45,7 @@ contract AaveTokenV3Harness is AaveTokenV3 {
   }
 
   function getDelegationState(address user) public view returns (uint8) {
-    return _balances[user].delegationMode;
+    return uint8(_balances[user].delegationMode);
   }
 
   function getNonce(address user) public view returns (uint256) {
